Remove old AFC comms code and migrate new implementation to afc.py
