--- src/harness/afc.cfg
@@ -1,3 +1,0 @@
-[AfcConfig]
-SpdAfcBaseUrl: 34.233.185.164/cgi-bin/test2.py
-AdminApiBaseUrl: 34.233.185.164/cgi-bin/test2.py
